Add schema helper
