+++ gems/smithy-cbor/spec/smithy-cbor/codec_spec.rb
@@ -5,8 +5,7 @@ require_relative '../spec_helper'
 module Smithy
   module CBOR
     describe Codec do
-      let(:sample_service) { ClientHelper.sample_service }
-      let(:service) { sample_service.const_get(:Schema).const_get(:SERVICE) }
+      let(:shape) { SchemaHelper.sample_schema.const_get(:Structure) }
 
       it 'serialize returns nil when given a unit shape' do
         expect(subject.serialize(Schema::Shapes::Prelude::Unit, '')).to be_nil
@@ -21,7 +20,6 @@ module Smithy
       end
 
       it 'serializes and deserializes data' do
-        shape = service.operation(:operation).input
         time = Time.now
         allow(Time).to receive(:at).and_return(time)
         data = {
@@ -53,14 +51,12 @@ module Smithy
 
       context 'structures' do
         it 'serializes and deserializes structures as a type' do
-          shape = service.operation(:operation).input
           type = shape.type.new(string: 'string')
           bytes = subject.serialize(shape, type)
           expect(subject.deserialize(shape, bytes).string).to eq('string')
         end
 
         it 'serializes and deserializes structures as a hash' do
-          shape = service.operation(:operation).input
           data = { string: 'string' }
           bytes = subject.serialize(shape, data)
           expect(subject.deserialize(shape, bytes).to_h).to eq(data)
@@ -69,7 +65,6 @@ module Smithy
 
       context 'unions' do
         it 'serializes and deserializes union as a type' do
-          shape = service.operation(:operation).input
           union = shape.member(:union).shape.member_type(:string).new('string')
           type = shape.type.new(union: union)
           bytes = subject.serialize(shape, type)
@@ -77,35 +72,30 @@ module Smithy
         end
 
         it 'serializes and deserializes union as a hash' do
-          shape = service.operation(:operation).input
           data = { union: { string: 'string' } }
           bytes = subject.serialize(shape, data)
           expect(subject.deserialize(shape, bytes).to_h).to eq(data)
         end
 
         it 'serializes a nil union' do
-          shape = service.operation(:operation).input
           data = { union: nil }
           bytes = subject.serialize(shape, data)
           expect(subject.deserialize(shape, bytes).union).to eq(nil)
         end
 
         it 'serializes an empty union' do
-          shape = service.operation(:operation).input
           data = { union: {} }
           bytes = subject.serialize(shape, data)
           expect(subject.deserialize(shape, bytes).union).to eq(nil)
         end
 
         it 'serializes nil union values' do
-          shape = service.operation(:operation).input
           data = { union: { string: nil } }
           bytes = subject.serialize(shape, data)
           expect(subject.deserialize(shape, bytes).to_h).to eq(data)
         end
 
         it 'deserializes unknown union members' do
-          shape = service.operation(:operation).input
           unknown_union_type = shape.member(:union).shape.member_type(:unknown)
           data = { union: { 'someThing' => 'someValue' } }
           deserialized = subject.deserialize(shape, CBOR.encode(data))
@@ -116,14 +106,12 @@ module Smithy
 
       context 'lists' do
         it 'serializes and deserializes lists' do
-          shape = service.operation(:operation).input
           data = { list: ['string'] }
           bytes = subject.serialize(shape, data)
           expect(subject.deserialize(shape, bytes).to_h).to eq(data)
         end
 
         it 'can handle sparse lists' do
-          shape = service.operation(:operation).input
           list_shape = shape.member(:list).shape
           list_shape.traits.merge!('smithy.api#sparse' => {})
           data = { list: [nil] }
@@ -134,14 +122,12 @@ module Smithy
 
       context 'maps' do
         it 'serializes and deserializes maps' do
-          shape = service.operation(:operation).input
           data = { map: { 'key' => 'value' } }
           bytes = subject.serialize(shape, data)
           expect(subject.deserialize(shape, bytes).to_h).to eq(data)
         end
 
         it 'can handle sparse maps' do
-          shape = service.operation(:operation).input
           map_shape = shape.member(:map).shape
           map_shape.traits.merge!('smithy.api#sparse' => {})
           data = { map: { 'key' => nil } }
