+++ gems/smithy-cbor/spec/spec_helper.rb
@@ -11,7 +11,7 @@ end
 require 'smithy-cbor'
 
 require 'smithy'
-require_relative '../../smithy-client/spec/support/client_helper'
+require_relative '../../smithy-schema/spec/support/schema_helper'
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
